webbeadando/kiado/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
chardet==3.0.4
Django==5.1.2
django-modeltranslation==0.19.11
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pillow==11.0.0
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.2
whitenoise==6.8.2
